prepearing to the test

+++ a%b.py
@@ -1,4 +1,4 @@
-a=int(input('a='))
+a=float(input('a='))
 b=int(input('b='))
 print('a%b =', a%b)
 print('b%a =', b%a)
